From procedural to object-oriented programming for building type objects
CityTiler code now uses object-oriented programming.

Next step consists in updating the documentation.
